fts: extract reader/filter.rs (ExcludeFilter, AtomExcludeFilter)
Phase 1c (2/3). Move the negation gates — the term ExcludeFilter and the
phrase-aware AtomExcludeFilter, both skip-probing negated cursors — into
reader/filter.rs. AtomExcludeFilter is pub(super); ExcludeFilter stays
pub(crate) (PreparedClauses carries it). Uses cursor/phrase/work through
their pub(super) surfaces. Pure move.
